requirements.txt
django>=1.5
python-statsd==1.5.8
